feat(功能): 实现回抽停止功能并优化系统架构
- 新增回抽停止功能，可通过配置回抽时间
- 重构系统架构，支持 OTA 升级和 Web 服务器
- 优化电机控制逻辑，增加忙状态处理
- 改进数据包处理流程，提高通信可靠性
- 更新文件系统操作，使用 LittleFS

+++ Arduino/include/Motor.h
@@ -4,8 +4,8 @@
 class Motor
 {
 private:
-    int pin1;
-    int pin2;
+    uint8_t pin1;
+    uint8_t pin2;
     bool isStop = true;
     String state = "停止";
     unsigned long startTime = 0;
@@ -13,7 +13,7 @@ private:
     bool banMotor = false;
 
 public:
-    Motor(int pin1, int pin2);
+    Motor(uint8_t pin1, uint8_t pin2);
     void forward(unsigned long);
     void backforward(unsigned long);
     void stop();
